Major: 0.0.90
New Features:
1. Make track number can be shown the whole track.
2. Make size alignment right.
3. Add some MediaInfo header for OS X.

+++ src/Modules/Music/Libraries/knmusictagid3v1.h
@@ -17,7 +17,7 @@ public:
         QString album;
         QString year;
         QString comment;
-        qint64 track;
+        quint8 track;
         QString genre;
     };
     explicit KNMusicTagID3v1(QObject *parent = 0);
